+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Unreleased
 * Fixes the issue when a technical review already exists so a new one is not created
 * [BREAKING] Updates the PHT endpoints such that they are coherent e,g `POST entity/create`
 * [BREAKING] Updates the retrieval of entities endpoint from `entity/list/{id}` to be `entity/users/{id}/entity`
+* Panel `auto-create` can now update the submitted proposals for `Science Verification`
+* Fixes the bug with the report endpoint.
+* Creates Science reviews as part of the panel update endpoint.
 
 3.1.1
 ***********